# Optimize contiguous sparse slicing