Significant speed up for contiguous slices

I had totally thought this was already implemented...

TODO:

- [x] Benchmarks

* deferred
  - ~[ ] Reverse slicing~
  - ~[ ] Steps other that 1?~

It's probably optimal to do this up to pretty large step sizes. So long as two adjacent rows are in the same slice we should probably take this approach.

cc: @ilan-gold

Benchmark:

```python
from anndata.experimental import write_elem, sparse_dataset
from scipy import sparse
import numpy as np
import zarr

z = zarr.open("test.zarr")
write_elem(
    z,
    "/",
    sparse.random(
        100_000,
        10_000,
        format="csr",
        random_state=np.random.default_rng(),
    )
)

x = sparse_dataset(z["/"])
%time x[10000:20000]
```

This PR:

```
CPU times: user 9.57 ms, sys: 5.09 ms, total: 14.7 ms
Wall time: 14.4 ms
<10000x10000 sparse matrix of type '<class 'numpy.float64'>'
	with 999208 stored elements in Compressed Sparse Row format>
```

Main:

```
CPU times: user 9.35 s, sys: 5.14 s, total: 14.5 s
Wall time: 14.8 s
<10000x10000 sparse matrix of type '<class 'numpy.float64'>'
	with 999208 stored elements in Compressed Sparse Row format>
```